Path fixing and refactor parser folder.
