+++ src/app/encryptedChatProtocol/commonTypes.ts
