+++ src/app/encryptedChatProtocol/requestPackets/NewToken.ts
@@ -1,6 +1,6 @@
-import { IBuilder } from "../../../common/IBuilder";
+import { IBuilder } from "../../common/IBuilder";
 import { PacketType } from "../commonTypes";
-import Packet from "../Packet";
+import Packet from "../../utils/parser/Packet";
 
 export default class NewTokenRequestPacket extends Packet {
     readonly refreshToken: string;
