+++ src/app/encryptedChatProtocol/requestPackets/Register.ts
@@ -1,6 +1,6 @@
-import { IBuilder } from "../../../common/IBuilder";
+import { IBuilder } from "../../common/IBuilder";
 import { AuthAttributs, PacketType } from "../commonTypes";
-import Packet from "../Packet";
+import Packet from "../../utils/parser/Packet";
 
 export default class RegisterRequstPacket extends Packet {
     readonly userAttributs: AuthAttributs;
