+++ src/app/encryptedChatProtocol/requestPackets/RequsetPacket.ts
@@ -1,5 +1,5 @@
 import { PacketType } from "../commonTypes";
-import Packet from "../Packet";
+import Packet from "../../utils/parser/Packet";
 
 export default abstract class RequestPacket extends Packet {
     constructor(type: PacketType, packetId?: string) {
