+++ src/app/encryptedChatProtocol/responsePackets/CreateChat.ts
@@ -1,5 +1,5 @@
 import { PacketType, Status } from "../commonTypes";
-import { IBuilder } from "../../../common/IBuilder";
+import { IBuilder } from "../../common/IBuilder";
 import ResponsePacket from "./ResponsePacket";
 
 export default class CreateChatResponsePacket extends ResponsePacket {
