+++ src/app/encryptedChatProtocol/responsePackets/NewToken.ts
@@ -1,5 +1,5 @@
 import { PacketType, Status } from "../commonTypes";
-import { IBuilder } from "../../../common/IBuilder";
+import { IBuilder } from "../../common/IBuilder";
 import ResponsePacket from "./ResponsePacket";
 
 export default class NewToken extends ResponsePacket {
