+++ src/app/encryptedChatProtocol/responsePackets/Register.ts
@@ -1,5 +1,5 @@
 import { PacketType, Status } from "../commonTypes";
-import { IBuilder } from "../../../common/IBuilder";
+import { IBuilder } from "../../common/IBuilder";
 import ResponsePacket from "./ResponsePacket";
 
 export default class RegisterResponsePacket extends ResponsePacket {
