+++ src/app/tasks/authentication/AuthRepository.ts
@@ -1,6 +1,6 @@
 import { IResult } from "../../../common/IResult";
 import UserRepository from "../../db/utils/UserRepository";
-import { AuthAttributs } from "../../utils/encryptedChatProtocol/commonTypes";
+import { AuthAttributs } from "../../encryptedChatProtocol/commonTypes";
 import { IAuthRepository } from "./IAuthRepository";
 import bcrypt from "bcrypt";
 import userAttributsSchema from "../../validations/packetValidations/schemas/userAttributsSchema";
