+++ src/app/tasks/register/index.ts
@@ -1,6 +1,6 @@
-import { PacketType, Status } from "../../utils/encryptedChatProtocol/commonTypes";
-import * as RequestPackets from "../../utils/encryptedChatProtocol/requestPackets";
-import * as ResponsePackets from "../../utils/encryptedChatProtocol/responsePackets";
+import { Status } from "../../encryptedChatProtocol/commonTypes";
+import * as RequestPackets from "../../encryptedChatProtocol/requestPackets";
+import * as ResponsePackets from "../../encryptedChatProtocol/responsePackets";
 import { IMessageSender } from "../../data/IMessageSender";
 import AuthRepository from "../authentication/AuthRepository";
