+++ src/app/tasks/token/index.ts
@@ -1,9 +1,9 @@
 import jwt from "jsonwebtoken";
 import env from "../../config/env.json";
 import { IResult } from "../../../common/IResult";
-import { PacketType, Status } from "../../utils/encryptedChatProtocol/commonTypes";
-import * as RequestPackets from "../../utils/encryptedChatProtocol/requestPackets";
-import * as ResponsePackets from "../../utils/encryptedChatProtocol/responsePackets";
+import { Status } from "../../encryptedChatProtocol/commonTypes";
+import * as RequestPackets from "../../encryptedChatProtocol/requestPackets";
+import * as ResponsePackets from "../../encryptedChatProtocol/responsePackets";
 import { IConnectedUserManeger } from "../../data/IConnectedUserMeneger";
 
 type UserSign = {
