+++ src/app/utils/parser/IHandler.ts
@@ -1,0 +1,3 @@
+import Packet from "./Packet";
+
+export type Handler<T> = (req: T, res: unknown/*IResponse*/) => void;
