+++ src/app/validations/packetValidations/response/register.ts
@@ -1,4 +1,4 @@
-import { AuthAttributs, PacketType, Status } from "../../../utils/encryptedChatProtocol/commonTypes";
+import { AuthAttributs, PacketType, Status } from "../../../encryptedChatProtocol/commonTypes";
 import { IResult } from "../../../../common/IResult";
 import { registerResponsePacketSchema } from "./schemas";
